% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvis.R
\docType{package}
\name{rvis}
\alias{rvis}
\title{rvis: This package is a tool-box for visualization in R and making word-class and elegant shiny dashboards.}
\section{Class DASHBOARD}{

rvis provides a Reference class named as DASHBOARD.
You can design your interactive dashboard layout and assign multiple R objects into your dashboard.
It supports various visualization engines like graphics, ggplot2, ggvis, googlevis, amCharts, rCharts, dygraphs and many other plotter engines.
}

